project/requirements.txt
torch>=1.12
torchvision
timm
torchmetrics
Pillow
numpy
scikit-learn